requirements_neuron.txt
absl-py==2.2.2
astor==0.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
dmlc-nnvm==1.19.6.0+0
dmlc-topi==1.19.6.0+0
dmlc-tvm==1.19.6.0+0
filelock==3.16.1
fsspec==2025.3.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.6.0
idna==3.10
importlib_metadata==8.5.0
inferentia-hwm==1.17.6.0+fbcd6c853
islpy==2023.1
Jinja2==3.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==2.6.3
neuron-cc==1.24.0.0+d58fa6134
numpy==1.22.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opt_einsum==3.4.0
pillow==10.4.0
protobuf==3.20.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1723018227672/work
requests==2.32.3
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tensorboard==1.15.0
tensorflow==1.15.5.post1
tensorflow-estimator==1.15.1
termcolor==2.4.0
torch==1.13.1+cpu
torch-neuron==1.13.1.2.11.13.0
torchvision==0.14.1+cpu
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
